profile/requirements.txt
Flask==0.10.1
flask-restful
flask-mysql
PyJWT
Pillow
